Fix main adapter and view holders
(cherry picked from commit bbd908393e7025f12d07cd9bd789004f0d837784)
