+++ app/src/main/java/jt/projects/perfectday/presentation/calendar/dateFragment/ChosenDateAdapter.kt
@@ -5,7 +5,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import jt.projects.model.DataModel
 import jt.projects.perfectday.databinding.ItemBirthdayFromPhoneBinding
-import jt.projects.perfectday.presentation.MainAdapter
+import jt.projects.perfectday.presentation.today.adapter.MainAdapter
 
 class ChosenDateAdapter : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
